.history/requirements_20250604100110.txt
customtkinter==5.2.0
PyPDF2==3.0.1
python-docx==0.8.11
arabic-reshaper==3.0.0
python-bidi==0.4.2
Pillow==10.2.0 